refactor(project-engine-client): move mock out of published files
The Counterfact mock lived under src/mock, so files:["src"] shipped it in the
published tarball as dead weight (never imported; counterfact stays a devDep).
Move it to a top-level mock/ dir outside the files allowlist: client consumers
now install only src/ (typed client + generated types), while the workspace /
e2e harness still boots it from source via `npm run mock` / `npm run test:e2e`.

- run.js packageRoot now climbs one level (mock/) not two (src/mock/)
- .nycrc scopes mock/**; test imports + e2e RUNNER path updated
- README/spike-doc path refs updated; added a 'Not published' note
- npm pack --dry-run: 8 files, zero mock/counterfact; 100% coverage holds
