+++ packages/spacecat-shared-project-engine-client/mock/counterfact/routes/v1/workspaces/{id}/projects.js
