+++ packages/spacecat-shared-project-engine-client/mock/stateful.js
